pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
target-version = ['py37']

[tool.coverage.run]
branch = true
source = ["footing"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "pass",
    "pytest.mark.skip"
]
show_missing = true
fail_under = 100

[tool.poetry]
name = "footing"
packages = [
  { include = "footing" }
]
exclude = [
  "*/tests/"
]
version = "0.1.7"
description = "Keep templated projects in sync with their template"
authors = ["Opus 10 Engineering"]
classifiers = [
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3 :: Only",
]
license = "BSD-3-Clause"
readme = "README.rst"
homepage = "https://github.com/Opus10/footing"
repository = "https://github.com/Opus10/footing"
documentation = "https://footing.readthedocs.io"

[tool.poetry.dependencies]
python = ">=3.7.0,<4"
importlib_metadata = { version = ">=4", python = "~3.7" }
click = ">=6.7"
cookiecutter = "<2.0.0"
formaldict = ">=1.0.5"
pexpect = ">=4.8.0"
pyyaml = ">=3.12"
python-gitlab = ">=2.10.1"
requests = ">=2.13.0"
shellingham = ">=1.5.0"
tldextract = ">=3.1.2"

[tool.poetry.dev-dependencies]
black = "22.3.0"
flake8 = "3.9.2"
flake8-bugbear = "22.1.11"
flake8-comprehensions = "3.8.0"
flake8-import-order = "0.18.1"
flake8-logging-format = "0.6.0"
flake8-mutable = "1.2.0"
git-tidy = "1.1.1"
myst-parser = "0.18.0"
pyfakefs = "3.1.0"
pytest = "6.2.5"
pytest-cov = "3.0.0"
pytest-mock = "3.7.0"
responses = "0.17.0"
Sphinx = "4.4.0"
sphinx-click = "3.0.3"
sphinx-rtd-theme = "1.0.0"
tox = "3.24.5"

[tool.poetry.scripts]
footing = 'footing.cli:main'

[tool.pytest.ini_options]
xfail_strict = true
testpaths = "footing/tests"
norecursedirs = ".venv"
